Fix typos, oauth-authorization-scope registration was already processed.

+++ apply-pending.sh
@@ -3,7 +3,11 @@ src="iana-ipp-registrations.xml"
 dst="-o ipp-registrations.xml"
 
 echo Fetching IANA source registry...
-curl -z $src -o $src http://www.iana.org/assignments/ipp-registrations/ipp-registrations.xml
+if test -f $src; then
+	curl -z $src -o $src http://www.iana.org/assignments/ipp-registrations/ipp-registrations.xml
+else
+	curl -o $src http://www.iana.org/assignments/ipp-registrations/ipp-registrations.xml
+fi
 
 if test $# -gt 0; then
 	files="$@"
@@ -16,7 +20,7 @@ for file in $files; do
 	title=`basename $file .txt | sed -e '1,$s/^vendor-//' | awk '{print toupper($1);}'`
 	url=`head -1 $file`
 	echo "    $file: $url"
-	./register $dst -x $url -t $title $src $file
+	./register $dst -x $url -t $title $src $file || exit 1
 	src="ipp-registrations.xml"
 	dst=""
 	echo ""
